+++ Frameworks/AVFoundation.framework/AVDispatchGroup.h
@@ -5,7 +5,7 @@
 @interface AVDispatchGroup : NSObject {
     NSObject<OS_dispatch_queue> * _ivarAccessQueue;
     NSArray * _notifyBlocks;
-    int  _outstandingOperations;
+    long long  _outstandingOperations;
 }
 
 - (void)dealloc;
