+++ Frameworks/AVFoundation.framework/AVMetadataObjectInternal.h
@@ -5,12 +5,12 @@
 @interface AVMetadataObjectInternal : NSObject {
     struct CGRect { 
         struct CGPoint { 
-            float x; 
-            float y; 
+            double x; 
+            double y; 
         } origin; 
         struct CGSize { 
-            float width; 
-            float height; 
+            double width; 
+            double height; 
         } size; 
     }  _bounds;
     struct { 
@@ -30,20 +30,20 @@
     NSString * _type;
 }
 
-@property struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; } bounds;
+@property struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; } bounds;
 @property struct { long long x1; int x2; unsigned int x3; long long x4; } duration;
 @property (retain) AVCaptureInput *input;
 @property (retain) AVMetadataObject *originalMetadataObject;
 @property struct { long long x1; int x2; unsigned int x3; long long x4; } time;
 @property (retain) NSString *type;
 
-- (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })bounds;
+- (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })bounds;
 - (void)dealloc;
 - (struct { long long x1; int x2; unsigned int x3; long long x4; })duration;
 - (id)init;
 - (id)input;
 - (id)originalMetadataObject;
-- (void)setBounds:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (void)setBounds:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
 - (void)setDuration:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
 - (void)setInput:(id)arg1;
 - (void)setOriginalMetadataObject:(id)arg1;
