+++ Frameworks/AVFoundation.framework/AVMutableVideoComposition.h
@@ -11,7 +11,7 @@
 @property (nonatomic) struct { long long x1; int x2; unsigned int x3; long long x4; } frameDuration;
 @property (nonatomic, copy) NSArray *instructions;
 @property (nonatomic) float renderScale;
-@property (nonatomic) struct CGSize { float x1; float x2; } renderSize;
+@property (nonatomic) struct CGSize { double x1; double x2; } renderSize;
 @property (nonatomic) int sourceTrackIDForFrameTiming;
 
 + (id)videoComposition;
@@ -28,7 +28,7 @@
 - (struct { long long x1; int x2; unsigned int x3; long long x4; })frameDuration;
 - (id)instructions;
 - (float)renderScale;
-- (struct CGSize { float x1; float x2; })renderSize;
+- (struct CGSize { double x1; double x2; })renderSize;
 - (void)setAnimationTool:(id)arg1;
 - (void)setBuiltInCompositorName:(id)arg1;
 - (void)setColorPrimaries:(id)arg1;
@@ -38,7 +38,7 @@
 - (void)setFrameDuration:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
 - (void)setInstructions:(id)arg1;
 - (void)setRenderScale:(float)arg1;
-- (void)setRenderSize:(struct CGSize { float x1; float x2; })arg1;
+- (void)setRenderSize:(struct CGSize { double x1; double x2; })arg1;
 - (void)setSourceTrackIDForFrameTiming:(int)arg1;
 - (int)sourceTrackIDForFrameTiming;
