+++ Frameworks/CloudKit.framework/CKPackageSection.h
@@ -3,20 +3,20 @@
  */
 
 @interface CKPackageSection : NSObject {
-    int  _index;
+    long long  _index;
     NSData * _signature;
     unsigned long long  _size;
 }
 
-@property (nonatomic, readonly) int index;
+@property (nonatomic, readonly) long long index;
 @property (nonatomic, readonly) NSData *signature;
 @property (nonatomic) unsigned long long size;
 
 - (void).cxx_destruct;
 - (id)CKPropertiesDescription;
 - (id)description;
-- (int)index;
-- (id)initWithIndex:(int)arg1 signature:(id)arg2 size:(unsigned long long)arg3;
+- (long long)index;
+- (id)initWithIndex:(long long)arg1 signature:(id)arg2 size:(unsigned long long)arg3;
 - (void)setSize:(unsigned long long)arg1;
 - (id)signature;
 - (unsigned long long)size;
