+++ Frameworks/ContactsUI.framework/CNStarkNoContentBannerView.h
@@ -13,7 +13,7 @@
 @property (nonatomic, retain) NSString *titleString;
 
 - (void).cxx_destruct;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
 - (void)setSubtitleLabel:(id)arg1;
 - (void)setSubtitleString:(id)arg1;
 - (void)setTitleLabel:(id)arg1;
