+++ Frameworks/ContactsUI.framework/CNStarkStyleProvider.h
@@ -4,16 +4,16 @@
 
 @interface CNStarkStyleProvider : CNStyleProvider
 
-+ (float)memberRowHeight;
++ (double)memberRowHeight;
 
 - (id)memberNameBoldFont;
 - (id)memberNameDisabledTextColor;
-- (float)memberNameFontSize;
+- (double)memberNameFontSize;
 - (id)memberNamePlaceholderFont;
 - (id)memberNameRegularFont;
 - (id)memberNameTextColor;
 - (id)membersBackgroundColor;
-- (unsigned int)membersCountToShowIndex;
+- (unsigned long long)membersCountToShowIndex;
 - (id)membersSearchBackgroundColor;
 - (id)membersSearchCellBackgroundColor;
 - (id)personLabelHighlightedColor;
