+++ Frameworks/ContactsUI.framework/CNTonePickerController.h
@@ -12,7 +12,7 @@
 - (void).cxx_destruct;
 - (void)cancelButton:(id)arg1;
 - (void)doneButton:(id)arg1;
-- (id)initWithAlertType:(int)arg1 activityAlert:(id)arg2;
+- (id)initWithAlertType:(long long)arg1 activityAlert:(id)arg2;
 - (void)setTonePicker:(id)arg1;
 - (id)tonePicker;
