+++ Frameworks/CoreAudioKit.framework/CAUITransportButton.h
@@ -7,12 +7,12 @@
     struct CGColor { } * fillColor;
     struct CGRect { 
         struct CGPoint { 
-            float x; 
-            float y; 
+            double x; 
+            double y; 
         } origin; 
         struct CGSize { 
-            float width; 
-            float height; 
+            double width; 
+            double height; 
         } size; 
     }  imageRect;
 }
@@ -26,7 +26,7 @@
 - (struct CGColor { }*)fillColor;
 - (void)flash;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
 - (struct CGPath { }*)newPathRefForStyle:(int)arg1;
 - (void)setDrawingStyle:(int)arg1;
 - (void)setFillColor:(struct CGColor { }*)arg1;
