+++ Frameworks/CoreData.framework/NSBatchUpdateResult.h
@@ -4,15 +4,15 @@
 
 @interface NSBatchUpdateResult : NSPersistentStoreResult {
     id  _aggregatedResult;
-    unsigned int  _resultType;
+    unsigned long long  _resultType;
 }
 
 @property (readonly) id result;
-@property (readonly) unsigned int resultType;
+@property (readonly) unsigned long long resultType;
 
 - (void)dealloc;
-- (id)initWithResult:(id)arg1 type:(unsigned int)arg2;
+- (id)initWithResult:(id)arg1 type:(unsigned long long)arg2;
 - (id)result;
-- (unsigned int)resultType;
+- (unsigned long long)resultType;
 
 @end
