+++ Frameworks/CoreData.framework/NSSQLOffsetIntermediate.h
@@ -3,10 +3,10 @@
  */
 
 @interface NSSQLOffsetIntermediate : NSSQLIntermediate {
-    unsigned int  _offset;
+    unsigned long long  _offset;
 }
 
 - (id)generateSQLStringInContext:(id)arg1;
-- (id)initWithOffset:(unsigned int)arg1 inScope:(id)arg2;
+- (id)initWithOffset:(unsigned long long)arg1 inScope:(id)arg2;
 
 @end
