+++ Frameworks/CoreData.framework/PFUbiquityGlobalObjectIDCache.h
@@ -16,7 +16,7 @@
 
 - (id)createGlobalIDForCompressedString:(id)arg1 withEntityNames:(id)arg2 primaryKeys:(id)arg3 peerIDs:(id)arg4;
 - (id)createGlobalIDForGlobalIDString:(id)arg1;
-- (id)createGlobalIDForPrimarKey:(unsigned int)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
+- (id)createGlobalIDForPrimarKey:(unsigned long long)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
 - (id)createGlobalIDForPrimaryKeyString:(id)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
 - (void)dealloc;
 - (id)init;
