+++ Frameworks/CoreFoundation.framework/_CFXNotificationDarwinTokenRegistration.h
@@ -8,7 +8,7 @@
 
 @property (readonly) int notifyToken;
 
-- (id)initWithToken:(unsigned long long)arg1 notifyToken:(int)arg2 options:(unsigned long)arg3 queue:(id)arg4 handler:(id /* block */)arg5;
+- (id)initWithToken:(unsigned long long)arg1 notifyToken:(int)arg2 options:(unsigned long long)arg3 queue:(id)arg4 handler:(id /* block */)arg5;
 - (void)invalidate;
 - (int)notifyToken;
