+++ Frameworks/CoreFoundation.framework/__NSOrderedSetSetProxy.h
@@ -7,7 +7,7 @@
 }
 
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
-- (unsigned int)count;
+- (unsigned long long)count;
 - (void)dealloc;
 - (id)initWithOrderedSet:(id)arg1;
 - (id)member:(id)arg1;
