+++ Frameworks/CoreImage.framework/CIPinchDistortion.h
@@ -18,7 +18,7 @@
 
 - (id)_pinchDistortionScaleGE1;
 - (id)_pinchDistortionScaleLT1;
-- (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })computeDOD:(float)arg1 scale:(float)arg2;
+- (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })computeDOD:(float)arg1 scale:(float)arg2;
 - (id)inputCenter;
 - (id)inputImage;
 - (id)inputRadius;
