+++ Frameworks/CoreImage.framework/CITextDetector.h
@@ -3,8 +3,8 @@
  */
 
 @interface CITextDetector : CIDetector {
-    float  _height;
-    float  _width;
+    double  _height;
+    double  _width;
     CIContext * context;
     NSMutableDictionary * featureOptions;
     FKTextDetector * textDetector;
@@ -12,9 +12,9 @@
 
 @property (nonatomic, retain) CIContext *context;
 
-- (id)adjustedImageFromImage:(id)arg1 orientation:(int)arg2 inverseCTM:(struct CGAffineTransform { float x1; float x2; float x3; float x4; float x5; float x6; }*)arg3;
+- (id)adjustedImageFromImage:(id)arg1 orientation:(int)arg2 inverseCTM:(struct CGAffineTransform { double x1; double x2; double x3; double x4; double x5; double x6; }*)arg3;
 - (id)context;
-- (struct CGAffineTransform { float x1; float x2; float x3; float x4; float x5; float x6; })ctmForImageWithBounds:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 orientation:(int)arg2;
+- (struct CGAffineTransform { double x1; double x2; double x3; double x4; double x5; double x6; })ctmForImageWithBounds:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1 orientation:(int)arg2;
 - (void)dealloc;
 - (id)featuresInImage:(id)arg1;
 - (id)featuresInImage:(id)arg1 options:(id)arg2;
