+++ Frameworks/CoreSpotlight.framework/CSPowerLog.h
@@ -3,27 +3,27 @@
  */
 
 @interface CSPowerLog : NSObject {
-    unsigned int  _cachedCount;
+    unsigned long long  _cachedCount;
     NSDate * _lastFlushDate;
     NSMutableDictionary * _operationsByBundleID;
 }
 
-@property (nonatomic) unsigned int cachedCount;
+@property (nonatomic) unsigned long long cachedCount;
 @property (nonatomic, retain) NSDate *lastFlushDate;
 @property (nonatomic, retain) NSMutableDictionary *operationsByBundleID;
 
-+ (id)keyNameForOperation:(int)arg1;
++ (id)keyNameForOperation:(long long)arg1;
 + (id)sharedInstance;
 
 - (void).cxx_destruct;
-- (void)_addToDictionary:(int)arg1 bundleID:(id)arg2 itemCount:(unsigned int)arg3;
-- (unsigned int)cachedCount;
+- (void)_addToDictionary:(long long)arg1 bundleID:(id)arg2 itemCount:(unsigned long long)arg3;
+- (unsigned long long)cachedCount;
 - (void)flushToPowerLog;
 - (id)init;
 - (id)lastFlushDate;
-- (void)logWithBundleID:(id)arg1 indexOperation:(int)arg2 itemCount:(unsigned int)arg3;
+- (void)logWithBundleID:(id)arg1 indexOperation:(long long)arg2 itemCount:(unsigned long long)arg3;
 - (id)operationsByBundleID;
-- (void)setCachedCount:(unsigned int)arg1;
+- (void)setCachedCount:(unsigned long long)arg1;
 - (void)setLastFlushDate:(id)arg1;
 - (void)setOperationsByBundleID:(id)arg1;
