+++ Frameworks/EventKitUI.framework/EKRecurrenceOrdinalPickerViewController.h
@@ -9,7 +9,7 @@
 @property (readonly, copy) NSString *debugDescription;
 @property <EKRecurrenceOrdinalPickerViewControllerDelegate> *delegate;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (nonatomic, readonly) UIPickerView *pickerView;
 @property (readonly) Class superclass;
 
@@ -18,12 +18,12 @@
 - (int)dayMask;
 - (id)delegate;
 - (void)loadView;
-- (int)numberOfComponentsInPickerView:(id)arg1;
+- (long long)numberOfComponentsInPickerView:(id)arg1;
 - (int)ordinalValue;
 - (id)pickerView;
-- (void)pickerView:(id)arg1 didSelectRow:(int)arg2 inComponent:(int)arg3;
-- (int)pickerView:(id)arg1 numberOfRowsInComponent:(int)arg2;
-- (id)pickerView:(id)arg1 titleForRow:(int)arg2 forComponent:(int)arg3;
+- (void)pickerView:(id)arg1 didSelectRow:(long long)arg2 inComponent:(long long)arg3;
+- (long long)pickerView:(id)arg1 numberOfRowsInComponent:(long long)arg2;
+- (id)pickerView:(id)arg1 titleForRow:(long long)arg2 forComponent:(long long)arg3;
 - (void)setDelegate:(id)arg1;
 - (void)updateFromRecurrenceRule:(id)arg1;
 - (void)updateRecurrenceRuleBuilder:(id)arg1;
